updates on MVVM
