+++ app/src/main/java/com/org/hospitalapplication/MainActivity.kt
@@ -23,28 +23,21 @@ import java.nio.charset.Charset
 class MainActivity : AppCompatActivity() {
 
     private lateinit var hospitalAdapter: HospitalAdapter
-    private lateinit var hospitalData:MutableList<HospitalData>
-
     private lateinit var hospitalViewModel:HospitalViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        hospitalData = readHospitalData()
 
-        hospitalViewModel = ViewModelProvider(this).get(HospitalViewModel::class.java)
 
-        hospitalViewModel.getHospitalList()
+        applicationContext.let {
+            hospitalViewModel = ViewModelProvider(this).get(HospitalViewModel::class.java)
 
-        hospitalViewModel.mHospitalDataList.observe(this, Observer {
-            hospitalAdapter =
-                HospitalAdapter(
-                    hospitalData
-                ) { hospitalData -> onHospitalClick(hospitalData) }
-            recycler_view_hospital_main.adapter = hospitalAdapter
-            recycler_view_hospital_main.layoutManager = LinearLayoutManager(this)
-        })
-    }
+        }
+        hospitalAdapter = HospitalAdapter(hospitalViewModel.getHospitalList()){ hospitalData -> onHospitalClick(hospitalData) }
+        recycler_view_hospital_main.adapter = hospitalAdapter
+        recycler_view_hospital_main.layoutManager = LinearLayoutManager(this)
+        }
 
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
@@ -70,50 +63,7 @@ class MainActivity : AppCompatActivity() {
         return true
     }
 
-    private fun readHospitalData() : MutableList<HospitalData>{
 
-        val hospitalList = ArrayList<HospitalData>()
-        val inputStream: InputStream = resources.openRawResource(R.raw.hospital)
-        val streamReader = InputStreamReader(inputStream, Charset.forName("UTF-8"))
-        val reader = BufferedReader(streamReader)
-        var line: String
-
-        try {
-            while (reader.readLine() != null){
-                line = reader.readLine().replace('\uFFFD',',')
-                val hospitalDetails =  line.split(",").toTypedArray()
-                val hospital = HospitalData(
-                    hospitalDetails[0].toInt(),
-                    hospitalDetails[1],
-                    hospitalDetails[2],
-                    hospitalDetails[3],
-                    hospitalDetails[4],
-                    hospitalDetails[5],
-                    hospitalDetails[6],
-                    hospitalDetails[7],
-                    hospitalDetails[8],
-                    hospitalDetails[9],
-                    hospitalDetails[10],
-                    hospitalDetails[11],
-                    hospitalDetails[12],
-                    hospitalDetails[13],
-                    hospitalDetails[14],
-                    hospitalDetails[15],
-                    hospitalDetails[16],
-                    hospitalDetails[17],
-                    hospitalDetails[18],
-                    hospitalDetails[19],
-                    hospitalDetails[20],
-                    hospitalDetails[21]
-                )
-                hospitalList.add(hospital)
-                Log.i("MyActivity", "$hospital")
-            }
-        } catch (e: RuntimeException) {
-            Log.e("MainActivity", "Error data was not able to read:" , e)
-        }
-        return hospitalList
-    }
 
     private fun onHospitalClick(detailPosition: HospitalData) {
